Moved the basic hash map to a collections package
