+++ pstatus-graphql-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/collections/BasicHashMap.kt
@@ -1,4 +1,4 @@
-package gov.cdc.ocio.processingstatusapi.plugins
+package gov.cdc.ocio.processingstatusapi.collections
 
 import kotlin.math.absoluteValue
